requirements.txt
torch>=1.9.0
transformers>=4.20.0
datasets>=2.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
requests>=2.25.0
pyyaml>=6.0
langdetect>=1.0.9
tqdm>=4.62.0
matplotlib>=3.5.0
seaborn>=0.11.0
